requirements.txt
Flask~=3.0.3
python-dotenv~=1.2.1
alembic~=1.17.0